requirements.txt
flask
google-generativeai
python-dotenv
pdf2image
